requirements.txt
Pillow
python-slugify
jinja2

# needs: exiftool
pyexiftool
